Move the vendored typing module into a vendor package
- makes a new package, vendor.python3 and moves typing.py into it

This is a bit neater, and helps resolve the import of the typing module in
_compat.py correctly in python2.7.

I.e, in python2.7, the lines

  from typing import *
  from typing import _Union

imports from the typing module in the same package, rather than
choosing the top-level typing module.  By making any such imports from
a vendor subpackage, we avoid this issue.