Vendor python2.7 typing.py
This ensures that all the latest features of typing.py (available in python 3.6)
are available; and allows the application of a small patch that fixes a major
issue.

Small Patch
-----------

The patch is on line #1223, where __subclasscheck__ is only permitted for
generic types from two modules, abc and functools.
```
    if sys._getframe(1).f_globals['__name__'] not in ['abc', 'functools']:
       ...
```

To support python2.7, we need to add 'singledispatch' to the list of permitted
modules, as singledispatch is packaged in it's own file singledispatch.py rather
than functools.py as it is in python3
```
    if sys._getframe(1).f_globals['__name__'] not in ['abc', 'functools',
                                                      'singledispatch']:
       ...
```