add optional namespace kws to pass into resolve_types
In the case of type annotations which are imported in a TYPE_CHECKING block
the origin class will not contain them in their runtime namespace so cattrs was
unable to resolve them which structuring and unstructuring.  Added a method to
register a custom namespace with the Converter class which could be used when
resolving classes not in the default namespace.

A simple test of nesting when those annotations of nested types are coming from
type checking blocks and therefore not present at runtime.  The test file is
test_deep_nest_custom_namespace.py