Add ForwardRef tests
These test all kinds of subtleties of ForwardRefs and their
(automatic, semiautomatic or manual) resolution. As dataclasses, attrs
classes and type aliases behave slightly different and, in addition to that,
behave differently with respect to PEP 563 (the delayed evaluation of type hints)
the tests are quite extensive. A critical test-case is also whether
resolution works across modules, i.e. whether ForwardRefs are handled correctly
when they are imported from another module.
Correct operation in any of the above cases cannot be taken for granted