Make validation notes picklable. (#408)
Certain test runners pickle the exceptions raised to accumulate them from
multiple processes.

For exceptions with `AttributeValidationNote` and/or `IterableValidationNote`
attached, because we have overridden `__new__` on these classes, pickle cannot
correctly unpickle them with the required arguments. For this use case,
exceptions being picklable is necessary to accumulate errors with the
`multiprocessing` setup.

This implements `__getnewargs__` to fix this problem.

Certain test runners that pickle the exceptions raised to accumulate
them from multiple processes. For exceptions with `AttributeValidationNote`
and/or `IterableValidationNote`, because we have overriden __new__ on
these classes, pickle is not able to correctly unpickle them with the required
arguments. This implements __getnewargs__ to fix this problem.

Signed-off-by: Zixuan James Li <p359101898@gmail.com>